feat: improve security

+++ api/v1/books/delete_book.js
@@ -4,6 +4,11 @@ module.exports = async (req, res) => {
   const id = parseInt(req.params.id);
 
   try {
+    const book = await db_books.getById(id);
+    if (!book) {
+      return res.status(404).json({ message: "Book not found." });
+    }
+    
     const deletedBook = await db_books.deleteById(id);
     return res.json({ message: "Livre supprimé", deleted: deletedBook[0] });
   } catch (err) {
